+++ tests/suite/performance_FO.ml
@@ -79,3 +79,14 @@ let () = declare "set"
   ~source_elpi:"set.elpi"
   ~description:"stdlib set"
   ()
+
+let () = declare "dt"
+  ~source_elpi:"dt.elpi"
+  ~description:"discrimination_tree indexing"
+  ()
+
+let () = declare "dt_off"
+  ~source_elpi:"dt_off.elpi"
+  ~description:"(without) discrimination_tree indexing"
+  ()
+